requirements.txt
# requirements.txt
tensorflow==2.15.0
scikit-learn==1.4.1
imbalanced-learn==0.12.0
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
google-cloud-storage==2.14.0